API,MAINT: cudf moved to the nicer `extract_datetime_component` API
Also change test a bit to ensure we cover all possible values in the
test.

This is a breaking change, as it removes some previously supported
things (day of year) and aligns with `pylibcudf`, which uses upper
case letters in the struct.
(Also renames the parameter, but it should be passed positionally.)

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>