ENH: Support more return columns
Each column takes up some space in the return and this space is
currently limited but settable.

This bumps the limit for all tasks so that we should be able to
use well above 100 columns everywhere.

(Clearly, it is very plausible that we might bump this further if
we run into a use-case that needs that.)

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>