ENH: Support `names=` in csv reading to not provide a header
We still don't allow usecols just being indices (we could).

In general, I would like to move to use csv_options at some point
(even if we have to copy out the ones we can deal with), but that
might take a bit and it seemed that this is a bit of a pain point.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>