DOC: Blocklist sphinx 8.2.0 for now
There are some new issues related to myst extension and also errors
importing a `pyi` file (because it tries to expand a default argument
that can't be expanded with just stub information?).

Maybe some of these need fixing, but on first sight it looks like it is
at least in part a regression that will hopefull fix itself by 8.2.1.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>