ENH: Implement basic reductions
This uses the simple-aggs/finalizer pattern internally used in cudf
(technically, we don't need the split as much here probably).

The implementation does not try to optimize the second step of the
reduction: it gathers all local results via a broadcast and then
does a single local second reduction.
In principle, a custom reduction (e.g. via legate) could optimize this,
although it is unclear if we lose much (since local results should be
small).

As of now, "count valid" is implemented as a special case, but it is
probably special enough that we should have a `count_valid()` (or just
`count_nulls()` task.
Once that exist, the special case here can be removed and instead that
task called directly.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>