API,BUG: Remove `global_num_rows`, use `is_partitioned`
This is breaking as we stop passing around `global_num_rows`, but this
should improve legate's ability to parallelize task launches (I think)
since fetching the `global_num_rows` on an unbound table requires blocking.

`is_partitioned` should also be a safer API here, because the local and
global number of rows could be equal if all but one chunks have 0 entries
(which is may be hard to trigger but is possible with filtering).

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>