API: Add `to_array()` methods to convert column or tables to arrays
I opted for `to_array()` because `to_cupynumeric()` just seems very
long, but happy to change!

It may be that, especially for columns, `cupynumeric.asarray(col)`
should just be the long-term solution.

For tables, I am personally not sure:  `pandas` does it, but in
practice implicitly converting columns into a new dimension isn't
that obvious to me.  Tables and arrays are very different.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>