Use arrow to read parquet metadata (#55)
This uses arrows parquet reader to read the metadata from the file. This
will need
a little bit of polishing for odder data types

A wrinkle is that the dtype conversion is certainly not complete (and
also note tested for all types here).

Should address the hopefully larger issue in gh-47, it will still
decompress too much data and use a larger memory overhead with large row
group chunk sizes.

(Not sure the CI will build, I think we effectively pull in arrow via
cudf, but we should have it as an explicit dependency with gh-53.)

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>
Signed-off-by: Rory Mitchell <r.a.mitchell.nz@gmail.com>