Add `LDF_PREFER_EAGER_ALLOCATIONS` for legate investigations
Legate folks asked a way to use eager allocations more.  This might
actually be useful for in some places (or at least the machinery around it).

I didn't set up the tests, I ran them locally with 2 GPUs and they run fine
except, that using this for joins is actually incorrect so one tests (correctly)
fails.

To use just run with `LDF_PREFER_EAGER_ALLOCATIONS=1`, and many ops will use
eager allocations:
* to/from timestamp
* casts
* unary and binary ops
* replace_nulls
* Join...
  * If LEFT join
  * and explicit right side broadcasting

(As said, even then it is incorrect: a left join can return a different
sized result if there are duplicates in the RHS.  Doing this will be
flagged as an error, and can lead to segfaults.)

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>