REV: Reinstate global memory resource (but only use it if GPUs are used)  (#58)
This brings back the global memory resource. We really need it since
libcudf
doesn't necessarily use the mr for all allocations and we want to cover
as
many as possible.
(I somewhat think mr() is really used only/mainly for memory allocations
returned.)

~Let's see if we can build the docs without a GPU runner with this. (If
this works, the CPU tests would work if they avoid all cupy calls/cudf
data creation, which they currently don't.)~

Docs build still requires a GPU runner for now. It seems that a legate
build for the GPU needs `libcuda.so` at runtime (quite likely something
else does to, but that is the first).

## Checklist
- [x] New or existing tests cover these changes.
- [x] The documentation is up to date with these changes.
- [x] Run `./build.sh test` for local testing, see
[CONTRIBUTING.md](https://github.com/rapidsai/legate-dataframe/blob/main/CONTRIBUTING.md#testing).
- [x] Ensure `pre-commit` checks are passing see
[CONTRIBUTING.md](https://github.com/rapidsai/legate-dataframe/blob/main/CONTRIBUTING.md#Pre-commit-hooks).

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>
Signed-off-by: Rory Mitchell <r.a.mitchell.nz@gmail.com>