MAINT,BUG: Indicate that we are using the task stream and side-effect (#67)
Based on guidance from the legate team, add the
`with_elide_device_ctx_sync` to indicate to realm that we are using the
task stream. This should still guarantee correct sychronization for
output tasks because sychronization still happens, we just indicate that
we do use the task stream and thus synching that is sufficient.

We also should indicate that the output csv/parquet writer tasks have
side effects. In practice, I don't think we hit any problems due to it,
but it is just correct (e.g. gh-64 is unrelated).

Closes gh-66

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>
Signed-off-by: Rory Mitchell <r.a.mitchell.nz@gmail.com>