BUG,API: Deal with NaNs->NULL explicitly inside the gpu task
I think we may want a nans_to_nulls utility anyway, but since we
currently have a CPU version and it deals with this in a well defined
way (I think), dealing with it inside the reduce code seemed reasonable
to me.

The up-side is that it is far more light-weight.  An interesting thing
might be groupby-aggs eventually, although for those nans_to_nulls may
not be that convenient either?

If cudf defines the behavior (in either direction), this should be revised probably.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>