Support logical operators (again) and in_between for polars
This re-adds supports for logical operators, which were missing because
arrow seems to require boolean inputs (i.e. an explicit cast).

Additionally, since `in_between` seems commonly used in polars ad
support for that again.
(Polars is a bit fun, in that it also only defines bitwise and/or it seems.)

With this, I can run pdsh query 6.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>