Add `LDF_PREFER_EAGER_ALLOCATIONS` for legate investigations (#57)
Legate folks asked a way to use eager allocations more. This might
actually be useful for in some places (or at least the machinery around
it).

I didn't set up the tests, I ran them locally with 2 GPUs and they run
fine except, that using this for joins is actually incorrect so one
tests (correctly) fails.

To use just run with `LDF_PREFER_EAGER_ALLOCATIONS=1`, and many ops will
use eager allocations:
* to/from timestamp
* casts
* unary and binary ops
* replace_nulls
* Join...
  * If LEFT join
  * and explicit right side broadcasting
* ~parquet reading~ (removed in this PR for now)

(As said, even then it is incorrect: a left join can return a different
sized result if there are duplicates in the RHS. Doing this will be
flagged as an error, and can lead to segfaults.)

---

I would be OK with putting this in if it helps for whatever tries
someone wants to do. I mean it if I say that the general ability may
actually be useful even if the environment variable is clearly wrong
with applying it to join.
It does add a bit of cruft, though.

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>
Signed-off-by: Rory Mitchell <r.a.mitchell.nz@gmail.com>