Add explicit NCCL depdencency
This is largely cargo-culted from raft.  It seems that newer legate
versions do not use nccl anymore and our transitive dependency will
break.

Not clearly tested without the new legate, so we may want to wait
until we actually see the change (which means until cupynumeric
updates it's legate dependency).

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>