Finish support for slice/head/tail in polars
In most cases, these are actually pushed to a lower level and would be
better tested on the individual function, but the test does catch
at least one of the table paths.

In practice, e.g. sorting also should get the `limit=` so that the
distributed sort can do less work (even if it still will do/need the final
slice).

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>