Expose datetime component to polars and fix true-divide (#100)
We should still expose datetime component extraction for CPU as well
probably, that might change things, but hopefully too annoyingly so.

I think true-divide (at least for polars) is main useful thing, while
divide I believe is a floor-divide, which we could expose but only
easily for non-floats probably.
So, hacked true-divide to work, although one day we may want to lower
that.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>