DOC: Add a note about polars support to docs
This adds a brief note about polars support to the docs.
Since (almost) all functionality is exposed as such, I think we
can do that.

The exact way of exposure should likely still change admittedly.
It would be nice to integrate as an `engine` like cudf-polars
does.

(I have not looked at it much yet, because, first, for the use-cases
that were discussed it seemed like we'll need the return of a `LogicalTable`
and, second, the CPU work got us further away of tightly integrating with
cudf-polars which would have been a good path.)

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>