Allow groupby count-valid with and without nulls and expose to polars
This channels through the FunctionOptions from arrow all the way into
the task, the nice thing is that those come already with serialization
capabilities.

The only thing really supported right now (on the GPU) is for counting
NULLs or not.  We could support more in the future.

We might add more options in the future (but maybe also disable CPU ones
if they are not ammendable to improving the distributed algorithm).

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>