Allow groupby count-valid with and without nulls and expose to polars (#116)
This channels through the FunctionOptions from arrow all the way into
the task, the nice thing is that those come already with serialization
capabilities.

The only thing really supported right now (on the GPU) is for counting
NULLs or not. We could support more in the future.

We might add more options in the future (but maybe also disable CPU ones
if they are not amendable to improving the distributed algorithm).

## Checklist
- [x] New or existing tests cover these changes.
- [x] The documentation is up to date with these changes.
- [x] Run `./build.sh test` for local testing, see
[CONTRIBUTING.md](https://github.com/rapidsai/legate-dataframe/blob/main/CONTRIBUTING.md#testing).
- [x] Ensure `pre-commit` checks are passing see
[CONTRIBUTING.md](https://github.com/rapidsai/legate-dataframe/blob/main/CONTRIBUTING.md#Pre-commit-hooks).

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>