Add column concatenate and use it for polars table concats
This adds column concatenation.  Honestly, it's a lot more tedious
then I expected this to be, because I don't think there is a convenient
way to concatenate arrays in legate yet.
(well at least string ones)

But this was one of the missing features and I started, so now it's
there, if maybe not ideal.
(I'll open a legate issue, I think it may be good to have some
abstraction for this type of thing in legion.)

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>