ENH: Add contains and polars is_in
Unfortunately, polars relies on list dtype here and we don't support
that yet, so there is a bit of a hack to rewrite the literal part
of the query to use the needed dtype and not the list one.
(I hope this is OK in practice and assume it can't really go wrong badly
although it can likely fail.)

Otherwise, relatively straight forward `contains` added in a `search.hpp`, etc.
the polars code is mostly cudf-polars boilerplate.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>