WIP: Allow `chunk_by_rows=True` for parquet reading
Allows to not read by row-group granularity and side-steps the
image constraints for the eager-alloc preference variable
(In python mode).
Option is always available though.

The new code fails for CPU strings (although only after the read is
done).  Something is invalid there, it isn't clear to me yet what
or why.
(Do recordbatches not slice gracefully for strings or so?)

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>