When easy coerce scalars via legate scalar -> store -> array
This acutely helps streaming workflows because the current scheme
disrupts the context.
I am not quite sure what the future would be, whether scalars should
be/need to be special, whether this should be moved to C++ (to support
all dtypes for one thing), or whether this may just go away again.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>