Use `large_string` always and change back list unpacking logic
Mixing string and `large_string` seemsm very annoying, since it breaks
`==` comparisons which would have to be carefull replaced everywhere.
Since I doubt there is a large downside to large strings and legate
technically is large string (64bit), decided to normalize to that for
now to side-step having to change the comparisons.

But... I could go either way, as it is it just created problems (to
be fair in part this might only work out because polars also uses
`large_string` always).

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>