Add basic distinct (unique) capability
This adds distinct (unique in polars/pandas, but libcudf call is
distinct and uses unique for the same operation on pre-sorted arrays).

Arrow seems not to have such a thing, one might be able to do it via
structured array columns as well, but groupby and unique are largely
the same operation, so I opted to implement it via groupby.

It seemed logical to optimize this for few results (i.e. many dropped
columns) by minimizing how much we shuffle, but added an option to
not do that.  (If legate-df ever does cardinality estimation via
metadata, that could be used for a heuristic here).

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>