Enable polars groupby min/max by adding minimal mask_nans functionality (#141)
This should be a bit nicer and it could be soon, but I also doubt that
it is speed relevant.
(I.e. the nice thing would be to just create a new mask, but that
requires the ability to create a new array that uses the old store
easily which should be in legate pretty soon though.)

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>