Do not check parquet directory for consistency
Maybe there is a better way to actually do some form of check?
The problem is that if we have multiple nodes/utility workers
then one of them may lag behind when the director already contains
files.
However, blocking on this also seems not better.  Is there a better
way that would allow checking then just an option to disable this
check?

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>